tests/requirements.txt
pytest
pytest-randomly
pytest-cov >=2.12.1
python-dateutil
types-python-dateutil
coverage !=6.0
